+++ lib/eve_online/esi/character_bookmark_folders.rb
@@ -1,0 +1,32 @@
+module EveOnline
+  module ESI
+    class CharacterBookmarkFolders < Base
+      API_ENDPOINT = 'https://esi.tech.ccp.is/v2/characters/%<character_id>s/bookmarks/folders/?datasource=tranquility'.freeze
+
+      attr_reader :character_id
+
+      def initialize(options)
+        super
+
+        @character_id = options[:character_id]
+      end
+
+      def bookmark_folders
+        output = []
+        response.each do |bookmark_folder|
+          output << Models::BookmarkFolder.new(bookmark_folder)
+        end
+        output
+      end
+      memoize :bookmark_folders
+
+      def scope
+        'esi-bookmarks.read_character_bookmarks.v1'
+      end
+
+      def url
+        format(API_ENDPOINT, character_id: character_id)
+      end
+    end
+  end
+end